codeclimate fix - reduce functions in TaskEntityService
