+++ src/main/java/uk/gov/crowncommercial/dts/scale/cat/processors/async/queueExecutor/TaskEntityService.java
@@ -21,8 +21,7 @@ import java.util.List;
 public class TaskEntityService {
     private final TaskRepo taskRepo;
     private final TaskGroupRepo taskGroupRepo;
-    private final EnvironmentConfig environmentConfig;
-    private final TaskRetryManager taskRetryManager;
+    private final TaskUtils taskUtils;
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void persist(String principal, Task task, String recordType, String recordId, String data) {
@@ -44,7 +43,7 @@ public class TaskEntityService {
         if(null != groupEntity.getNode())
             entity.setNode(groupEntity.getNode());
         else
-            entity.setNode(environmentConfig.getServiceInstance());
+            entity.setNode(taskUtils.getNodeName());
 
         entity.setTimestamps(Timestamps.createTimestamps(principal));
         Timestamps.updateTimestamps(entity.getTimestamps(), principal);
@@ -60,7 +59,7 @@ public class TaskEntityService {
         entity.setName(recordType);
         entity.setReference(recordId);
         entity.setStatus('A');
-        entity.setNode(environmentConfig.getServiceInstance());
+        entity.setNode(taskUtils.getNodeName());
         entity.setTimestamps(Timestamps.createTimestamps(principal));
         taskGroupRepo.saveAndFlush(entity);
         return entity;
@@ -142,7 +141,7 @@ public class TaskEntityService {
 
     private void updateHistory(TaskEntity entity, char taskExecutionStatus, String response) {
         TaskHistoryEntity history = getLatestHistory(entity);
-        if(!isClosed(history)) {
+        if(!taskUtils.isClosed(history)) {
             history.setStatus(taskExecutionStatus);
             history.setResponse(response);
             Timestamps.updateTimestamps(history.getTimestamps(), entity.getPrincipal());
@@ -161,10 +160,7 @@ public class TaskEntityService {
         }
     }
 
-    private boolean isClosed(TaskHistoryEntity entity){
-        char status = entity.getStatus();
-        return (status == 'C' || status == 'F' || status == 'A');
-    }
+
 
 
     private void addHistory(TaskEntity entity) {
@@ -213,19 +209,11 @@ public class TaskEntityService {
 
         int i = 0;
         while (null == entity && i < 6) {
-            sleep(1000);
+            taskUtils.sleep(1000);
             i++;
             entity = taskRepo.findById(task.getId()).orElse(null);
         }
         return entity;
-    }
-
-
-    private void sleep(int i) {
-        try {
-            Thread.sleep(i);
-        } catch (InterruptedException e) {
-        }
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
@@ -240,7 +228,7 @@ public class TaskEntityService {
 
     public boolean isSchedulable(Task task){
         TaskEntity entity = getEntity(task);
-        return isSchedulable(entity);
+        return taskUtils.isSchedulable(entity);
     }
 
     private void markTaskStatus(Task task, String response, char status) {
@@ -252,21 +240,13 @@ public class TaskEntityService {
         taskRepo.save(entity);
     }
 
-    public boolean isSchedulable(TaskEntity entity){
-        if(entity.getStatus() == Task.SCHEDULED &&
-                (null == entity.getNode() || entity.getNode().equalsIgnoreCase(environmentConfig.getServiceInstance()))){
-            return true;
-        }
-        return false;
-    }
-
     public TaskEntity assignToSelf(TaskEntity taskEntity) {
         TaskEntity entity = taskRepo.findById(taskEntity.getId()).orElseThrow();
-        entity.setNode(environmentConfig.getServiceInstance());
+        entity.setNode(taskUtils.getNodeName());
         entity.getTimestamps().setUpdatedAt(Instant.now());
         TaskGroupEntity groupEntity =  entity.getGroup();
         if(null != groupEntity){
-            groupEntity.setNode(environmentConfig.getServiceInstance());
+            groupEntity.setNode(taskUtils.getNodeName());
         }
         taskRepo.saveAndFlush(entity);
         return entity;
